src/test/java/com/bradsbrain/simpleastronomy/BaseUtilsTest.java
/*
 *  Copyright 2011 Brad Parks
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package com.bradsbrain.simpleastronomy;

import org.junit.Test;

import static org.hamcrest.MatcherAssert.assertThat;
import static org.hamcrest.Matchers.closeTo;

public class BaseUtilsTest {

    @Test
    public void testAdjustTo360Range() {
        assertThat(BaseUtils.adjustTo360Range(400), closeTo(40, 0.5));
        assertThat(BaseUtils.adjustTo360Range(-50), closeTo(310, 0.5));
    }
}
